Split stale review classification policy
