+++ src/supervisor/stale-review-bot-classification-policy.ts
@@ -1,3 +1,41 @@
+import type { GitHubPullRequest, IssueRunRecord, PullRequestCheck, ReviewThread, SupervisorConfig } from "../core/types";
+import { configuredReviewProviderKinds } from "../core/review-providers";
+import { hasProcessedReviewThread } from "../review-handling";
+import { allCodexConnectorRepairResidueThreadsAreP2 } from "../codex-connector-review-repair-coverage";
+import {
+  codexConnectorMustFixReviewThreads,
+  evaluateCodexConnectorConvergencePolicy,
+} from "../codex-connector-review-policy";
+import { clusterConfiguredBotReviewThreads } from "../codex-connector-review-churn";
+import {
+  configuredBotReviewFollowUpState,
+  configuredBotReviewThreads,
+  manualReviewThreads,
+  nonActionableConfiguredBotReviewThreads,
+  pendingBotReviewThreads,
+} from "../review-thread-reporting";
+import {
+  currentHeadRepairProofThreadFingerprint,
+  hasFreshCurrentHeadCodexSuccessReviewedCommit,
+  projectCurrentHeadCodexRepairProof,
+} from "../current-head-codex-repair-proof";
+import {
+  buildCurrentHeadCleanCommentResidueEvidence,
+  buildCurrentHeadCodexNoMajorSignalEvidence,
+  buildCurrentHeadVerificationEvidenceSummary,
+  buildCurrentHeadVerifiedRepairResidueArtifactEvidenceSummary,
+  hasCurrentHeadCodexTurnVerification,
+  hasCurrentHeadLocalCiVerification,
+  hasCurrentHeadMarkedNoSourceChangeCodexTurnVerification,
+  hasCurrentHeadNoSourceChangeCodexTurnVerification,
+  hasCurrentHeadSuccessSignal,
+} from "./stale-review-current-head-evidence";
+import {
+  deterministicRepositoryPathRepairProbeEvidence,
+  requiresDeterministicRepositoryPathRepairProbeEvidence,
+  type RepositoryFileContents,
+} from "./stale-review-repository-path-repair-evidence";
+
 export type StaleReviewBotClassificationOutcome =
   | "actionable_current_diff"
   | "metadata_only"
@@ -89,6 +127,27 @@ export interface StaleReviewBotAutoRepairSuppressionPolicyArgs {
   verifiedStaleResidue: boolean;
   actionableClusterCount: number;
   verifiedAutoResolveEnabled: boolean;
+}
+
+export interface StaleReviewBotRemediationClassificationArgs {
+  config: SupervisorConfig | null;
+  record: IssueRunRecord;
+  pr: GitHubPullRequest | null;
+  checks: PullRequestCheck[];
+  reviewThreads: ReviewThread[];
+  repositoryFileContents?: RepositoryFileContents;
+}
+
+export interface StaleReviewBotAutoRepairSuppressionClassificationArgs {
+  config: SupervisorConfig | null;
+  record: IssueRunRecord;
+  pr: GitHubPullRequest | null;
+  checks: PullRequestCheck[];
+  reviewThreads: ReviewThread[];
+  classification: StaleReviewBotClassificationOutcome;
+  missingProbeReason: string | null;
+  actionableMustFixThreads: ReviewThread[];
+  repeatStopExhausted: boolean;
 }
 
 function unresolvedWork(): StaleReviewBotClassificationPolicyDecision {
@@ -257,6 +316,252 @@ export function classifyStaleReviewBotRemediationPolicy(
   return args.provider === "codex" ? classifyCodexReviewBotPolicy(args) : classifyConfiguredBotPolicy(args);
 }
 
+function allChecksPassing(checks: Pick<PullRequestCheck, "bucket">[]): boolean {
+  return checks.length > 0 && checks.every((check) => check.bucket === "pass" || check.bucket === "skipping");
+}
+
+function hasFailingChecks(checks: Pick<PullRequestCheck, "bucket">[]): boolean {
+  return checks.some((check) => check.bucket === "fail");
+}
+
+function hasPendingChecks(checks: Pick<PullRequestCheck, "bucket">[]): boolean {
+  return checks.some((check) => check.bucket === "pending" || check.bucket === "cancel");
+}
+
+function hasCleanMergeState(pr: GitHubPullRequest): boolean {
+  return pr.state === "OPEN" && !pr.isDraft && pr.mergeStateStatus === "CLEAN" && pr.mergeable === "MERGEABLE";
+}
+
+function hasMergeConflictState(pr: GitHubPullRequest): boolean {
+  return pr.state !== "OPEN" || pr.isDraft || pr.mergeStateStatus === "DIRTY" || pr.mergeable === "CONFLICTING";
+}
+
+function hasProcessedCurrentHeadRepairProofReviewThread(
+  config: SupervisorConfig,
+  record: Pick<
+    IssueRunRecord,
+    "processed_review_thread_ids" | "processed_review_thread_fingerprints" | "last_head_sha"
+  >,
+  pr: Pick<GitHubPullRequest, "headRefOid">,
+  thread: ReviewThread,
+): boolean {
+  if (hasProcessedReviewThread(record, pr, thread)) {
+    return true;
+  }
+
+  const repairProofFingerprint = currentHeadRepairProofThreadFingerprint(config, pr, thread);
+  return repairProofFingerprint
+    ? hasProcessedReviewThread(record, pr, thread, repairProofFingerprint)
+    : false;
+}
+
+function classifyCodexMetadataOnly(
+  args: StaleReviewBotRemediationClassificationArgs & {
+    config: SupervisorConfig;
+    pr: GitHubPullRequest;
+  },
+): StaleReviewBotClassificationPolicyDecision {
+  const configuredThreads = configuredBotReviewThreads(args.config, args.reviewThreads);
+  const currentConfiguredThreads = configuredThreads.filter((thread) => !thread.isOutdated);
+  const policy = evaluateCodexConnectorConvergencePolicy(args.config, args.pr, args.reviewThreads);
+  const mustFixReviewThreads = codexConnectorMustFixReviewThreads(args.reviewThreads);
+  const currentHeadCleanCommentResidueEvidence = buildCurrentHeadCleanCommentResidueEvidence({
+    config: args.config,
+    record: args.record,
+    pr: args.pr,
+    reviewThreads: args.reviewThreads,
+    currentConfiguredThreads,
+    mustFixReviewThreads,
+  });
+  const hasMarkedNoSourceChangeRepair = hasCurrentHeadMarkedNoSourceChangeCodexTurnVerification(
+    args.record,
+    args.pr,
+  );
+  const hasCurrentHeadCodexTurnRepairVerification = hasCurrentHeadCodexTurnVerification(args.record, args.pr);
+  const checkEvidenceCanProveRepair = args.record.repair_attempt_count > 0;
+  const verificationEvidenceSummary = buildCurrentHeadVerificationEvidenceSummary({
+    config: args.config,
+    record: args.record,
+    pr: args.pr,
+    checks: args.checks,
+    allowCheckEvidence: checkEvidenceCanProveRepair,
+  });
+  const verifiedRepairArtifactEvidenceSummary =
+    buildCurrentHeadVerifiedRepairResidueArtifactEvidenceSummary(args);
+  const hasUnprocessedMustFix = mustFixReviewThreads.some((thread) =>
+    !hasProcessedCurrentHeadRepairProofReviewThread(args.config, args.record, args.pr, thread)
+  );
+  const unprocessedMustFixCanUseRepairProof =
+    allCodexConnectorRepairResidueThreadsAreP2(mustFixReviewThreads) &&
+    projectCurrentHeadCodexRepairProof(args)?.source === "finding_set_verification_artifact";
+  if (
+    verifiedRepairArtifactEvidenceSummary &&
+    (!hasUnprocessedMustFix || unprocessedMustFixCanUseRepairProof)
+  ) {
+    return {
+      classification: "verified_current_head_repair_pending_thread_resolution",
+      summary: VERIFIED_CURRENT_HEAD_REPAIR_SUMMARY,
+      verificationEvidenceSummary: verifiedRepairArtifactEvidenceSummary,
+    };
+  }
+
+  return classifyStaleReviewBotRemediationPolicy({
+    provider: "codex",
+    configuredThreadCount: configuredThreads.length,
+    currentConfiguredThreadCount: currentConfiguredThreads.length,
+    manualThreadCount: manualReviewThreads(args.config, args.reviewThreads).length,
+    sameHead: args.record.last_head_sha === args.pr.headRefOid,
+    allChecksPassing: allChecksPassing(args.checks),
+    cleanMergeState: hasCleanMergeState(args.pr),
+    mergeConflictState: hasMergeConflictState(args.pr),
+    pendingBotThreadCount: pendingBotReviewThreads(args.config, args.record, args.pr, currentConfiguredThreads).length,
+    followUpState: configuredBotReviewFollowUpState(args.config, args.record, args.pr, currentConfiguredThreads),
+    allCurrentConfiguredThreadsProcessed: currentConfiguredThreads.every((thread) =>
+      hasProcessedCurrentHeadRepairProofReviewThread(args.config, args.record, args.pr, thread),
+    ),
+    convergenceOutcome: policy?.outcome ?? null,
+    hasUnprocessedMustFix,
+    verificationEvidenceSummary,
+    noMajorSignalEvidence: buildCurrentHeadCodexNoMajorSignalEvidence({
+      record: args.record,
+      pr: args.pr,
+      reviewThreads: args.reviewThreads,
+      currentConfiguredThreads,
+    }),
+    currentHeadCleanCommentResidueEvidence,
+    deterministicProbeEvidence: deterministicRepositoryPathRepairProbeEvidence({
+      reviewThreads: args.reviewThreads,
+      repositoryFileContents: args.repositoryFileContents,
+    }),
+    hasMarkedNoSourceChangeRepair,
+    verifiedNoSourceChangeRepair: hasCurrentHeadNoSourceChangeCodexTurnVerification(
+      args.record,
+      args.pr,
+      mustFixReviewThreads,
+    ),
+    hasExplicitCurrentHeadRepairVerification: hasCurrentHeadCodexTurnRepairVerification,
+    hasCurrentHeadRepairCheckVerification:
+      !hasMarkedNoSourceChangeRepair && hasCurrentHeadLocalCiVerification(args.record, args.pr),
+    repairAttemptCount: args.record.repair_attempt_count,
+    allMustFixRepairResidueThreadsAreP2: allCodexConnectorRepairResidueThreadsAreP2(mustFixReviewThreads),
+    requiresDeterministicRepairProbeEvidence: requiresDeterministicRepositoryPathRepairProbeEvidence(args.reviewThreads),
+    currentHeadSuccess: hasCurrentHeadSuccessSignal(args.pr),
+  });
+}
+
+export function classifyStaleReviewBotRemediation(
+  args: StaleReviewBotRemediationClassificationArgs,
+): StaleReviewBotClassificationPolicyDecision {
+  if (!args.config || !args.pr) {
+    return classifyStaleReviewBotRemediationPolicy({
+      provider: "configured_bot",
+      configuredThreadCount: 0,
+      currentConfiguredThreadCount: 0,
+      manualThreadCount: 0,
+      sameHead: false,
+      allChecksPassing: false,
+      cleanMergeState: false,
+      mergeConflictState: false,
+      pendingBotThreadCount: 0,
+      followUpState: "inactive",
+      allCurrentConfiguredThreadsProcessed: false,
+      convergenceOutcome: null,
+      hasUnprocessedMustFix: false,
+      verificationEvidenceSummary: null,
+      noMajorSignalEvidence: null,
+      currentHeadCleanCommentResidueEvidence: null,
+      deterministicProbeEvidence: null,
+      hasMarkedNoSourceChangeRepair: false,
+      verifiedNoSourceChangeRepair: false,
+      hasExplicitCurrentHeadRepairVerification: false,
+      hasCurrentHeadRepairCheckVerification: false,
+      repairAttemptCount: 0,
+      allMustFixRepairResidueThreadsAreP2: false,
+      requiresDeterministicRepairProbeEvidence: false,
+      currentHeadSuccess: false,
+    });
+  }
+
+  const { config, record, pr, checks, reviewThreads } = args;
+  const configuredThreads = configuredBotReviewThreads(config, reviewThreads);
+  if (configuredReviewProviderKinds(config).includes("codex")) {
+    return classifyCodexMetadataOnly({
+      config,
+      record,
+      pr,
+      checks,
+      reviewThreads,
+      repositoryFileContents: args.repositoryFileContents,
+    });
+  }
+
+  return classifyStaleReviewBotRemediationPolicy({
+    provider: "configured_bot",
+    configuredThreadCount: configuredThreads.length,
+    currentConfiguredThreadCount: configuredThreads.length,
+    manualThreadCount: manualReviewThreads(config, reviewThreads).length,
+    sameHead: record.last_head_sha === pr.headRefOid,
+    allChecksPassing: allChecksPassing(checks),
+    cleanMergeState: hasCleanMergeState(pr),
+    mergeConflictState: hasMergeConflictState(pr),
+    pendingBotThreadCount: pendingBotReviewThreads(config, record, pr, configuredThreads).length,
+    followUpState: configuredBotReviewFollowUpState(config, record, pr, configuredThreads),
+    allCurrentConfiguredThreadsProcessed: configuredThreads.every((thread) => hasProcessedReviewThread(record, pr, thread)),
+    convergenceOutcome: null,
+    hasUnprocessedMustFix: false,
+    verificationEvidenceSummary: null,
+    noMajorSignalEvidence: null,
+    currentHeadCleanCommentResidueEvidence: null,
+    deterministicProbeEvidence: null,
+    hasMarkedNoSourceChangeRepair: false,
+    verifiedNoSourceChangeRepair: false,
+    hasExplicitCurrentHeadRepairVerification: false,
+    hasCurrentHeadRepairCheckVerification: false,
+    repairAttemptCount: record.repair_attempt_count,
+    allMustFixRepairResidueThreadsAreP2: false,
+    requiresDeterministicRepairProbeEvidence: false,
+    currentHeadSuccess: Boolean(pr.configuredBotCurrentHeadObservedAt && pr.configuredBotCurrentHeadStatusState === "SUCCESS"),
+  });
+}
+
+export function isProvenStaleReviewMetadataClassification(
+  classification: StaleReviewBotClassificationOutcome,
+): boolean {
+  return (
+    classification === "metadata_only" ||
+    classification === "metadata_only_current_head_converged" ||
+    classification === "verified_no_source_change_pending_thread_resolution" ||
+    classification === "verified_current_head_repair_pending_thread_resolution"
+  );
+}
+
+export function isVerifiedStaleResidueClassification(
+  classification: StaleReviewBotClassificationOutcome,
+): boolean {
+  return (
+    classification === "verified_no_source_change_pending_thread_resolution" ||
+    classification === "verified_current_head_repair_pending_thread_resolution"
+  );
+}
+
+export function isPolicyResolvableStaleReviewBotClassification(
+  classification: StaleReviewBotClassificationOutcome,
+): boolean {
+  return classification === "metadata_only" || classification === "metadata_only_current_head_converged";
+}
+
+export function verifiedStaleReviewResidueAutoResolveEnabled(
+  config: SupervisorConfig,
+  classification: StaleReviewBotClassificationOutcome,
+): boolean {
+  return (
+    (classification === "verified_no_source_change_pending_thread_resolution" &&
+      config.verifiedNoSourceChangeReviewThreadAutoResolve === true) ||
+    (classification === "verified_current_head_repair_pending_thread_resolution" &&
+      config.verifiedCurrentHeadRepairReviewThreadAutoResolve === true)
+  );
+}
+
 export function classifyStaleReviewBotAutoRepairSuppressionPolicy(
   args: StaleReviewBotAutoRepairSuppressionPolicyArgs,
 ): StaleReviewBotAutoRepairSuppressedReason {
@@ -289,4 +594,59 @@ export function classifyStaleReviewBotAutoRepairSuppressionPolicy(
   }
 
   return "none";
+}
+
+export function classifyStaleReviewBotAutoRepairSuppression(
+  args: StaleReviewBotAutoRepairSuppressionClassificationArgs,
+): StaleReviewBotAutoRepairSuppressedReason {
+  const { config, pr, checks } = args;
+  return classifyStaleReviewBotAutoRepairSuppressionPolicy({
+    hasConfigAndPr: Boolean(config && pr),
+    repeatStopExhausted: args.repeatStopExhausted,
+    manualOrUnconfiguredReviewThreads: Boolean(
+      config &&
+        (manualReviewThreads(config, args.reviewThreads).length > 0 ||
+          nonActionableConfiguredBotReviewThreads(config, args.reviewThreads).length > 0),
+    ),
+    mergeConflictState: Boolean(pr && hasMergeConflictState(pr)),
+    failingChecks: hasFailingChecks(checks),
+    pendingChecks: hasPendingChecks(checks),
+    missingProbeReason: args.missingProbeReason,
+    verifiedStaleResidue: isVerifiedStaleResidueClassification(args.classification),
+    actionableClusterCount: clusterConfiguredBotReviewThreads(args.actionableMustFixThreads).length,
+    verifiedAutoResolveEnabled: Boolean(config && verifiedStaleReviewResidueAutoResolveEnabled(config, args.classification)),
+  });
+}
+
+export function codexConnectorCurrentHeadReviewState(args: {
+  config: SupervisorConfig | null;
+  record: IssueRunRecord;
+  pr: GitHubPullRequest;
+  reviewThreads: ReviewThread[];
+}): "observed" | "requested" | "missing" | "not_applicable" {
+  if (!args.config || !configuredReviewProviderKinds(args.config).includes("codex")) {
+    return "not_applicable";
+  }
+
+  if (hasFreshCurrentHeadCodexSuccessReviewedCommit(args.pr, args.reviewThreads)) {
+    return "observed";
+  }
+
+  if (
+    args.pr.configuredBotCurrentHeadObservationSource === "codex_pr_success_comment" &&
+    args.pr.configuredBotCurrentHeadObservedAt
+  ) {
+    return "observed";
+  }
+
+  const recordRequestedSha = args.record.codex_connector_review_requested_head_sha;
+  const prRequestedSha = args.pr.codexConnectorReviewRequestedHeadSha;
+  if (
+    (args.record.codex_connector_review_requested_observed_at || args.pr.codexConnectorReviewRequestedAt) &&
+    (recordRequestedSha ?? prRequestedSha) === args.pr.headRefOid
+  ) {
+    return "requested";
+  }
+
+  return "missing";
 }
